# swagger.json
{"openapi":"3.0.0","paths":{"/conversations/{id}":{"get":{"operationId":"ConversationController_getConversationById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found conversation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Conversation"}}}}},"tags":["conversations"]}},"/conversations/user/{userId}":{"get":{"operationId":"ConversationController_getConversationOfUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found conversations","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Conversation"}}}}}},"tags":["conversations"]}},"/conversations/user/{userId1}/receiver/{userId2}":{"get":{"operationId":"ConversationController_getConversationBetweenUsers","parameters":[{"name":"userId1","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId2","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found conversation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Conversation"}}}},"201":{"description":"The created conversation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Conversation"}}}}},"tags":["conversations"]}},"/conversations":{"post":{"operationId":"ConversationController_createConversation","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConversationRequest"}}}},"responses":{"201":{"description":"The created conversation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Conversation"}}}},"500":{"description":"Conversation already exists"}},"tags":["conversations"]}},"/conversations/{conversationId}/messages":{"get":{"operationId":"MessageController_readAllMessages","parameters":[{"name":"conversationId","required":true,"in":"path","schema":{"type":"string"}},{"name":"readerId","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"offset","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"The found messages of conversation with id <conversationId>. \n\t\t\t\t\t  Will mark as read all messages from conversation.\n\t\t\t\t\t  You can use query string limit and offset for retrieving:\n\t\t\t\t\t\te.g. limit = 1 and no offset will read last message","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Message"}}}}}},"tags":["messages"]},"post":{"operationId":"MessageController_sendMessage","parameters":[{"name":"conversationId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageRequest"}}}},"responses":{"201":{"description":"The created message","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}}},"tags":["messages"]}},"/conversations/{conversationId}/messages/{messageId}":{"get":{"operationId":"MessageController_readMessageById","parameters":[{"name":"conversationId","required":true,"in":"path","schema":{"type":"string"}},{"name":"messageId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The found message","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}}},"tags":["messages"]}},"/conversations/{conversationId}/messages/{messageId}/markAsRead":{"post":{"operationId":"MessageController_markAsReadMessageById","parameters":[{"name":"conversationId","required":true,"in":"path","schema":{"type":"string"}},{"name":"messageId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The result of operation","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["messages"]}}},"info":{"title":"Chat Service","description":"The Chat Service API description","version":"1.0","contact":{}},"tags":[{"name":"Chat Service","description":""}],"servers":[],"components":{"schemas":{"Message":{"type":"object","properties":{"_id":{"type":"string"},"conversationId":{"type":"string"},"senderId":{"type":"string"},"receiverId":{"type":"string"},"sentDate":{"format":"date-time","type":"string"},"isRead":{"type":"boolean"},"seenDate":{"format":"date-time","type":"string"},"content":{"type":"string"}},"required":["_id","conversationId","senderId","receiverId","sentDate","isRead","seenDate","content"]},"Conversation":{"type":"object","properties":{"_id":{"type":"string"},"participants":{"type":"array","items":{"type":"string"}},"lastMessage":{"$ref":"#/components/schemas/Message"},"messages":{"type":"array","items":{"type":"string"}}},"required":["_id","participants","lastMessage","messages"]},"ConversationRequest":{"type":"object","properties":{"participants":{"type":"array","items":{"type":"string"}}},"required":["participants"]},"MessageRequest":{"type":"object","properties":{"senderId":{"type":"string"},"receiverId":{"type":"string"},"content":{"type":"string"}},"required":["senderId","receiverId","content"]}}}}